+++ Vungle/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
   * 6.3.17.0
     * This version of the adapters has been certified with Vungle 6.13.17.
-    * Deprecate the use of placement IDs at Vungle's initialization time per Vungle's new SDK.
+    * Remove the placement IDS (pids) parameter from Vungle's init call per Vungle's new SDK requirements.
     * Update the implementation of Vungle's `updateConsentStatus()` to now also pass a `consentMessageVersion`.
 
   * 6.2.5.1